move lms2012 firmware headers to 'contrib' folder
