+++ contrib/lms2012/ev3_analog.h
